[rcr] Re-export useMemoCache in top level React namespace
In order to support using the compiler on versions of React prior to 19,
we need the ability to statically import `c` (aka useMemoCache) or
fallback to a polyfill supplied by `react-compiler-runtime` (note: this
is a separate npm package, not to be confused with
`react/compiler-runtime`, which is currently a part of react).

To do this we first need to re-export `useMemoCache` under the top level
React namespace again, which is additive and thus non-breaking. Doing so
allows `react-compiler-runtime` to statically either re-export `React.c`
or supply a polyfill, without the need for a dynamic import which is
finicky to support due to returning a promise.

In later PRs I will remove `react/compiler-runtime` and update the
compiler to emit imports to `React.c` instead.

ghstack-source-id: 70352fc9e1f9411297fc56eabb23d40c95cec1b9
Pull Request resolved: https://github.com/facebook/react/pull/31050